feat: cleaned up beginner template
