+++ python/create-onchain-agent/templates/beginner/.gitignore
