+++ python/create-onchain-agent/templates/beginner/pyproject.toml.jinja
