+++ python/create-onchain-agent/templates/chatbot/prepare_agentkit.py.jinja
@@ -28,7 +28,34 @@ from eth_account import Account
 from coinbase_agentkit.network import NETWORK_ID_TO_CHAIN_ID, NETWORK_ID_TO_CHAIN, CHAIN_ID_TO_NETWORK_ID
 {% endif %}
 
-load_dotenv()
+"""
+AgentKit Configuration
+
+This file serves as the entry point for configuring AgentKit tools and wallet providers.
+It handles wallet setup, persistence, and initializes AgentKit with the appropriate providers.
+
+# Key Steps to Configure AgentKit:
+
+1. Set up your WalletProvider:
+   - Learn more: https://github.com/coinbase/agentkit/tree/main/python/agentkit#evm-wallet-providers
+
+2. Set up your Action Providers:
+   - Action Providers define what your agent can do.  
+   - Choose from built-in providers or create your own:
+     - Built-in: https://github.com/coinbase/agentkit/tree/main/python/coinbase-agentkit#create-an-agentkit-instance-with-specified-action-providers
+     - Custom: https://github.com/coinbase/agentkit/tree/main/python/coinbase-agentkit#creating-an-action-provider
+
+# Next Steps:
+
+- Explore the AgentKit README: https://github.com/coinbase/agentkit
+- Learn more about available WalletProviders & Action Providers.
+- Experiment with custom Action Providers for your unique use case.
+
+## Want to contribute?
+Join us in shaping AgentKit! Check out the contribution guide:  
+- https://github.com/coinbase/agentkit/blob/main/CONTRIBUTING.md
+- https://discord.gg/CDP
+"""
 
 # Configure a file to persist wallet data
 wallet_data_file = "wallet_data.txt"
@@ -156,12 +183,9 @@ def prepare_agentkit():
     agentkit = AgentKit(AgentKitConfig(
         wallet_provider=wallet_provider,
         action_providers=[
-            {% if _wallet_provider == "cdp" %}
-            cdp_wallet_action_provider(),
-            {% endif %}
+            {% if _wallet_provider == "cdp" %}cdp_wallet_action_provider(),{% endif %}
             cdp_api_action_provider(),
             erc20_action_provider(),
-            pyth_action_provider(),
             wallet_action_provider(),
             weth_action_provider(),
         ]
